Modified the front page.
I have no idea where I'm going with this (Actually, I do). The goal is
to show a bit more information about myself as I'm a freelancer and
nowhere on the site does it mention it. Also, I wanted to reduce the
height of the header.

Also, I want it to be easier to browse the published post so I made
thumbnails that includes excerpt so someone can right the first sentence
of every post and see if it might interest them.

Also moved the social stuff at the bottom as I'm not sure it's relevant.

It's very rough but still I'm going to publish this as is as I don't
want to have massive changes that won't make it due to stupid reason
(doubts).

So publish now, iterate later.