DRY the form builder
The post and partial editor use pretty much the same editor. The only
difference is the publish/draft status that post have. For that reason,
it's much better to have a class that is reused to generate both editor.

PostBuilder & PartialBuilder are now subclasses of Admin::EditorBuilder.
There may be additional refactor possible but I think it's a good
stepping stone.