Fixed CSS spaghetti and added a new section
Right now, the CSS for the recipe function is hardcoded even though I
would prefer to use a partial for this. The reason why I'm not using a
partial is that partials can't handle custom data.

Any idea would be to add support for dynamic data like:

<link rel="partials" href="/partials/1"
data-recipe='http://lamarmite.ca/recipes/1'>

And then use something like moustache syntax in the partial.

a href="{{recipe}}

food for thoughts.